Fix parentheses for UNION and SELECT...AS
Previously sub SELECTS were always wrapped in parentheses which resulted
in invalid SQL syntax for UNIONs for sqlite3 and for SELECT...AS for
MySQL as reported by Issue #167.

Now SelectStmts are type cast to a builder type that hides the
SelectStmt type from interpolate().

Additionally all tests were improved by adding subtests for each SQL
driver. This allows tests for all drivers to continue to be run even if
a test for a single driver fails.