Add quoteindent for select table name
Makes select consistent with the other query builders for the table naming in generated queries. 

Before this, running a e2e lifecycle test where select, insert, update, and delete are all called for the same table/resource/relation in a postgres db, the select statement would fail because the relation could not be found. The other queries would all succeed.